mscng: avoid double BCryptDestroyKey() on pubkey without cert
This is not needed, we already destroy the pubkey a few lines above. I
somehow remembered about CRYPT_ACQUIRE_CACHE_FLAG is used for the pubkey
handle (so we would have a non-owning reference to the pubkey in case it
comes from a cert), but 1) that's a flag for
CryptAcquireCertificatePrivateKey() 2) we don't use it even there.