MSCNG: raw DSA support (#158)
* mscng: initial support for public keys without a certificate (DSA case)

- support reading a DSA pubkey directly from XML
- support adopt/duplicate/finalize operations for key data with a pubkey only
  (without a certificate)

Testcase: merlin-xmldsig-twenty-three/signature-enveloped-dsa (only partially passes)

* mscng: support writing public keys without a certificate (DSA case)

Testcase: merlin-xmldsig-twenty-three/signature-enveloped-dsa (now fully passes)

* mscng: avoid double BCryptDestroyKey() on pubkey without cert

This is not needed, we already destroy the pubkey a few lines above. I
somehow remembered about CRYPT_ACQUIRE_CACHE_FLAG is used for the pubkey
handle (so we would have a non-owning reference to the pubkey in case it
comes from a cert), but 1) that's a flag for
CryptAcquireCertificatePrivateKey() 2) we don't use it even there.