nss: fix assert condition when getting key type of ECDSA key
The condition is now consistent with the RSA getter. In practice this
fixes a crash when using libxmlsec via its API and setting only the
private key (but not the public key) for signing -- as
SECKEY_GetPublicKeyType(NULL) is not safe.

Bugreport: https://bugs.documentfoundation.org/show_bug.cgi?id=109180