mscng: add keysstore support
This also requires changes in the x509 key data class. It's no longer safe to
call CertDeleteCertificateFromStore() for no longer needed certificates, as the
store may be a non-in-memory one, and the user expects that the system store is
only read by xmlsec. Non-deleting also means that we can't assert anymore that
the in-memory cert store is empty when we delete it.

We don't leak memory with this, as the refcount of the certificate contexts is
still properly decremented.

As a side effect 6 TODOs added earlier are now resolved.