configure: macOS improvement (#197)
* mscng, x509 verify: implement recursive check for untrusted certs

This is the case when the issuer certificate is not trusted, parents are not in
the trusted store, but we can find a parent in the untrusted store which has a
parent in the trusted store.

Testcase: XMLSEC_TEST_NAME="merlin-xmldsig-twenty-three/signature" nmake check-dsig

* mscng: implement loading .der keys

Testcase: XMLSEC_TEST_NAME="signature-big" nmake check-dsig

* xmlenc: fix non-matching function name in xmlSecError() calls

* xmldsig: fix non-matching function name in xmlSecError() calls

* configure: fix shared library suffix on macOS

See <https://lists.gnu.org/archive/html/libtool/2010-04/msg00020.html>,
shrext_cmds is a command to be evaluated. Linux uses .so for both
loadable and shared libraries (in terms of libtool), so the problem is
not visible there, but on macOS this caused configure to literally look
for

libnspr4`test .$module = .yes && echo .so || echo .dylib`

for the nss backend, while nss provides libnspr4.dylib.