tests: add coverage for --privkey-der outside the gcrypt backend (#235)
* tests: add coverage for --privkey-der outside the gcrypt backend

Everything else only used --pkcs12 for testing, but at least openssl has
a working --privkey-der implementation. Test that as well.

* Disable the new enveloping-sha256-rsa-sha256-der test on Windows

My intention with the test was to cover something that already works, so
for now just don't run this new test on Windows. Needs further
investigation.

* Address review comments

Remove the cert from the test (and x509 from enabled data) to ensure
that key file is actually used for both signature and verification.