build: do not extern "C" scope for include files
C++ include files such as icu and others do not expect the scope to be in C,
as include directive is expected to be performed at global scope.

Example[1].

Yes, and include file of C++ can have extern "C++" to force C++ scope,
however, it is expected that global scope will be C++ already and most headers
do not do that.

[1] https://github.com/unicode-org/icu/pull/572

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>