Fix 'make clean' before first full build
When the build is configured out of tree (srcdir != builddir)
clean-local in docs/api tried to chmod a number of files and directories
prior to removing them. However make has not yet run these files and
directories won't exist yet and chmod will exit with a non-zero exit status
which aborts the make clean command with an error.
The changed build rules will do the file permission changes as part of the
find command, which will just skip over non-existing directories and files
instead of generating errors.