Fix PKCS12_parse return code handling (#310)
In OpenSSL, PKCS12_parse returns 0 in case of error. This commit
properly handle the error case.

Note: it is not a severe error since if an error occurs when parsing the
pkcs12 keystore, the call to xmlSecOpenSSLEvpKeyAdopt will immediatly
fail. But it will not report the real root cause in the logs:

It will report:
  func=xmlSecOpenSSLEvpKeyAdopt:file=evp.c:line=351:obj=unknown:subj=pKey != NULL:error=100:assertion:

With the fix, we will have the following line in the logs:
  func=xmlSecOpenSSLAppPkcs12LoadBIO:file=app.c:line=633:obj=unknown:subj=PKCS12_parse:error=4:crypto library function failed:openssl error: 50856204: digital envelope routines: NULL unsupported